jpeg decoder cleanup

+++ lib/src/formats/jpeg/jpeg_component.dart
@@ -3,7 +3,7 @@ part of image;
 class JpegComponent {
   int h;
   int v;
-  final List<Int32List> quantizationTableList;
+  final List<Int16List> quantizationTableList;
   int quantizationIndex;
   int blocksPerLine;
   int blocksPerColumn;
@@ -15,6 +15,6 @@ class JpegComponent {
   JpegComponent(this.h, this.v, this.quantizationTableList,
       this.quantizationIndex);
 
-  Int32List get quantizationTable =>
+  Int16List get quantizationTable =>
       quantizationTableList[quantizationIndex];
 }
